JTest cleanup
- Parameters given sensible defaults
- Classes/filenames renamed to express purpose rather than base class
- JANA1 noise removed
- Compiler warnings fixed
