+++ src/plugins/JTest/JTestPlotter.h
@@ -5,17 +5,16 @@
 #include <JANA/JEventProcessor.h>
 #include <JANA/JEvent.h>
 
-class JTestEventProcessor : public JEventProcessor {
+class JTestPlotter : public JEventProcessor {
 
-private:
-    size_t m_cputime_ms = 10;
-    size_t m_write_bytes = 2000000;
+    size_t m_cputime_ms = 20;
+    size_t m_write_bytes = 1000;
     double m_cputime_spread = 0.25;
     double m_write_spread = 0.25;
 
 public:
 
-    JTestEventProcessor(JApplication* app) : JEventProcessor(app) {
+    JTestPlotter(JApplication* app) : JEventProcessor(app) {
 
         auto params = japp->GetJParameterManager();
 
@@ -25,20 +24,11 @@ public:
         params->SetDefaultParameter("jtest:plotter_spread", m_cputime_spread, "");
     }
 
-    ~JTestEventProcessor() {}
-
-    const char* className() {
-        return "JTestEventProcessor";
-    }
-
-    void Init() {
-    }
-
     void Process(const std::shared_ptr<const JEvent>& aEvent) {
 
         // Read the track data
         auto td = aEvent->GetSingle<JTestTrackData>();
-        auto sum = read_memory(td->buffer);
+        read_memory(td->buffer);
 
         // Consume CPU
         consume_cpu_ms(m_cputime_ms, m_cputime_spread);
@@ -48,8 +38,6 @@ public:
         write_memory(hd->buffer, m_write_bytes, m_write_spread);
         aEvent->Insert(hd);
     }
-
-    void Finish() {}
 
 };
